Feature/#90: 다섯 번째 온보딩 뷰 구현 및 공통 컴포넌트 분리
